app/requirements.txt
anyio==4.7.0
asgiref==3.8.1
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
pyparsing==3.2.0
python-dotenv==1.0.1
python-telegram-bot==21.8
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
